+++ src/library/Cell.js
@@ -1,0 +1,6 @@
+import styled from "styled-components";
+
+const Cell = styled.div`
+`
+
+export { Cell }
