+++ src/library/CollapseButton.js
@@ -1,0 +1,23 @@
+import styled from "styled-components"
+
+const CollapseButton = styled.div`
+  padding: 15px 0;
+
+  color: rgba(0,0,0,.5);
+
+  text-align: center;
+  text-transform: uppercase;
+
+  line-height: 1.6em;
+  letter-spacing: 1px;
+
+  font-size: 16px;
+  font-weight: 400;
+  font-style: normal;
+
+  border-bottom-style: solid;
+  border-bottom-color: rgba(0,0,0,.1);
+  border-bottom-width: 2px;
+`
+
+export { CollapseButton }
