+++ src/library/Footer.js
@@ -1,0 +1,11 @@
+import styled from "styled-components";
+
+const Footer = styled.div`
+  padding: 50px 0;
+
+  border-top-style: solid;
+  border-top-color: rgba(0,0,0,.1);
+  border-top-width: 2px;
+`
+
+export { Footer }
