+++ src/library/Grid.js
@@ -1,0 +1,8 @@
+import styled from "styled-components";
+
+const Grid = styled.div`
+  display: grid;
+  padding: 0 1rem;
+`
+
+export { Grid }
