+++ src/library/Hero.js
@@ -1,0 +1,8 @@
+import styled from "styled-components";
+
+const Hero = styled.div`
+  position: relative;
+`
+
+
+export { Hero }
