+++ src/library/HeroTitle.js
@@ -1,0 +1,17 @@
+import styled from "styled-components";
+
+const HeroTitle = styled.div`
+  position: absolute;
+  top: 50%;
+  left: 0;
+  right: 0;
+  z-index: 1;
+
+  font-size: 21px;
+
+  color: white;
+  text-transform: uppercase;
+  text-align: center;
+`
+
+export { HeroTitle }
