+++ src/library/Icon.js
@@ -1,0 +1,9 @@
+import styled from "styled-components";
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
+
+const Icon = styled(FontAwesomeIcon)`
+  color: black;
+  font-size: 24px;
+`
+
+export { Icon }
