+++ src/library/Navbar.js
@@ -1,0 +1,16 @@
+import styled from "styled-components";
+
+const Navbar = styled.div`
+  display: grid;
+
+  padding: 1rem 0;
+
+  text-align: center;
+  text-transform: uppercase;
+  
+  border-bottom-style: solid;
+  border-bottom-color: rgba(0,0,0,.1);
+  border-bottom-width: 2px;
+`
+
+export { Navbar }
