+++ src/library/NavbarItem.js
@@ -1,0 +1,18 @@
+import styled from "styled-components";
+
+const NavbarItem = styled.div`
+  padding: 0.3rem 0;
+
+  color: rgba(0,0,0,.2);
+
+  text-transform: uppercase;
+
+  line-height: 1.6em;
+  letter-spacing: 1px;
+
+  font-size: 16px;
+  font-weight: 400;
+  font-style: normal;
+`
+
+export { NavbarItem }
