+++ src/library/Section.js
@@ -1,0 +1,9 @@
+import styled from "styled-components";
+
+const Section = styled.div`
+  margin: 40px 0;
+
+  text-align: center;
+`
+
+export { Section }
