+++ src/library/SectionBody.js
@@ -1,0 +1,14 @@
+import styled from "styled-components";
+
+const SectionBody = styled.p`
+  text-align: left;
+  font-size: 14px;
+  line-height: 1.8em;
+  text-transform: none;
+  text-decoration: none;
+  font-weight: 300;
+  font-style: normal;
+  margin: 0 0 1.8em;
+`
+
+export { SectionBody }
