+++ src/library/SectionHeader.js
@@ -1,0 +1,14 @@
+import styled from "styled-components";
+
+const SectionHeader = styled.h3`
+  margin: 1em 0 .3em;
+  font-size: 1.1em;
+  line-height: 1.5em;
+  text-transform: none;
+  text-decoration: none;
+  letter-spacing: 1px;
+  font-weight: 400;
+  font-style: normal;
+`
+
+export { SectionHeader }
