+++ src/pages/index.js
@@ -1,184 +1,28 @@
 import "../styles/reboot.css"
 
+import {
+  Grid,
+  Cell,
+  Navbar,
+  NavbarItem,
+  ListGroup,
+  ListGroupItem,
+  Collapse,
+  CollapseButton,
+  Hero,
+  HeroTitle,
+  Section,
+  SectionHeader,
+  SectionBody,
+  Icon,
+  Footer,
+} from "../library"
+
 import React from 'react'
-import { Link } from 'gatsby'
+
 import Img from 'gatsby-image'
-
-import styled, { createGlobalStyle, css } from "styled-components";
-import { CSSTransition } from 'react-transition-group';
-
-import { SlideDown } from 'react-slidedown'
-import 'react-slidedown/lib/slidedown.css'
-
-
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
+import { Link } from 'gatsby'
 import { faGithub, faLinkedin, faTwitter } from '@fortawesome/free-brands-svg-icons'
-
-const GlobalStyle = createGlobalStyle`
-  html,
-  body,
-  #___gatsby,
-  #___gatsby > div {
-    width: 100%;
-    height: 100%;
-  }
-`
-
-const Grid = styled.div`
-  display: grid;
-  padding: 0 1rem;
-`
-
-const Cell = styled.div`
-`
-
-const Navbar = styled.div`
-  display: grid;
-
-  padding: 1rem 0;
-
-  text-align: center;
-  text-transform: uppercase;
-  
-  border-bottom-style: solid;
-  border-bottom-color: rgba(0,0,0,.1);
-  border-bottom-width: 2px;
-`
-
-const NavbarItem = styled.div`
-  padding: 0.3rem 0;
-
-  color: rgba(0,0,0,.2);
-
-  text-transform: uppercase;
-
-  line-height: 1.6em;
-  letter-spacing: 1px;
-
-  font-size: 16px;
-  font-weight: 400;
-  font-style: normal;
-`
-
-const NavbarButton = styled.div`
-  padding: 15px 0;
-
-  color: rgba(0,0,0,.5);
-
-  text-align: center;
-  text-transform: uppercase;
-
-  line-height: 1.6em;
-  letter-spacing: 1px;
-
-  font-size: 16px;
-  font-weight: 400;
-  font-style: normal;
-
-  border-bottom-style: solid;
-  border-bottom-color: rgba(0,0,0,.1);
-  border-bottom-width: 2px;
-`
-
-const Header = styled.div`
-  padding: 30px 0;
-  text-align: center;
-`
-
-const ListGroup = styled.div`
-  width: 100%;
-
-  padding: 15px 0;
-
-  color: rgba(0,0,0,.3);
-  line-height: 2em;
-  text-transform: uppercase;
-  text-decoration: none;
-  letter-spacing: 3px;
-
-  font-size: 14px;
-  font-weight: 600;
-  font-style: normal;
-
-  text-align: center;
-
-  ${({ horizontal }) => horizontal && css`
-    display: flex;
-    flex-direction: row;
-  `}
-`
-
-const ListGroupItem = styled.div`
-
-  ${({ horizontal }) => horizontal && css`
-    padding: 0 4px;
-  `}
-`
-
-const Hero = styled.div`
-  position: relative;
-`
-
-const HeroTitle = styled.div`
-  position: absolute;
-  top: 50%;
-  left: 0;
-  right: 0;
-  z-index: 1;
-
-  font-size: 21px;
-
-  color: white;
-  text-transform: uppercase;
-  text-align: center;
-`
-
-const Section = styled.div`
-  text-align: center;
-  margin: 40px 0;
-`
-
-const SectionBody = styled.p`
-  text-align: left;
-  font-size: 14px;
-  line-height: 1.8em;
-  text-transform: none;
-  text-decoration: none;
-  font-weight: 300;
-  font-style: normal;
-  margin: 0 0 1.8em;
-`
-
-const SectionHeader = styled.h3`
-  margin: 1em 0 .3em;
-  font-size: 1.1em;
-  line-height: 1.5em;
-  text-transform: none;
-  text-decoration: none;
-  letter-spacing: 1px;
-  font-weight: 400;
-  font-style: normal;
-`
-
-const Footer = styled.div`
-  padding: 50px 0;
-
-  border-top-style: solid;
-  border-top-color: rgba(0,0,0,.1);
-  border-top-width: 2px;
-`
-
-const Collapse = styled(SlideDown)`
-
-  &.react-slidedown.gm-slidedown {
-    transition-duration: 150ms;
-    transition-timing-function: ease-in-out;
-  }
-`
-const SocialIcon = styled(FontAwesomeIcon)`
-  color: black;
-  font-size: 24px;
-`
 
 
 export default class IndexPage extends React.Component {
@@ -212,17 +56,17 @@ export default class IndexPage extends React.Component {
                 ) : null
               }
             </Collapse>
-            <NavbarButton onClick={this.handleMenuClick}>menu</NavbarButton>
+            <CollapseButton onClick={this.handleMenuClick}>menu</CollapseButton>
           </Cell>
           <Cell>
-            <Header>
+            <Section>
               <Img fixed={this.props.data.logo.fixed} />
               <ListGroup>
                 <ListGroupItem>51 edgemere street</ListGroupItem>
                 <ListGroupItem>montauk, ny, 11954</ListGroupItem>
                 <ListGroupItem>(800) 123-4567</ListGroupItem>
               </ListGroup>
-            </Header>
+            </Section>
           </Cell>
           <Cell>
             <Hero>
@@ -252,16 +96,15 @@ export default class IndexPage extends React.Component {
           <Cell>
             <Footer>
                 Made by Gatsby Manor
-
               <ListGroup horizontal>
                 <ListGroupItem horizontal>
-                  <SocialIcon icon={faGithub} />
+                  <Icon icon={faGithub} />
                 </ListGroupItem>
                 <ListGroupItem horizontal>
-                  <SocialIcon icon={faLinkedin} />
+                  <Icon icon={faLinkedin} />
                 </ListGroupItem>
                 <ListGroupItem horizontal>
-                  <SocialIcon icon={faTwitter} />
+                  <Icon icon={faTwitter} />
                 </ListGroupItem>
               </ListGroup>
             </Footer>
